+++ src/main/java/com/example/miniproject/dto/response/MemberResponseDto.java
@@ -1,25 +1,21 @@
-package com.example.miniproject.controller.response;
+package com.example.miniproject.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+
 import java.time.LocalDateTime;
 
-@Builder
+
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
-@Getter
-public class PostResponseDto {
-
+@Builder
+public class MemberResponseDto {
     private Long id;
+    private String userId;
     private String nickname;
-    private String productUrl;
-    private String productName;
-    private Integer star;
-    private String content;
-    private String imageUrl;
-    private Long likes;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
 }
