+++ src/main/java/com/example/miniproject/dto/response/PostResponseDto.java
@@ -1,21 +1,25 @@
-package com.example.miniproject.dto;
+package com.example.miniproject.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
 import java.time.LocalDateTime;
 
-
-@Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-@Builder
-public class MemberResponseDto {
+@Getter
+public class PostResponseDto {
+
     private Long id;
-    private String userId;
     private String nickname;
+    private String productUrl;
+    private String productName;
+    private Integer star;
+    private String content;
+    private String imageUrl;
+    private Long likes;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
 }
