+++ src/main/java/com/example/miniproject/repository/PostRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
     List<Post> findAllByOrderByModifiedAtDesc();
-    List<Post> findAllByMember(Member member);
+    List<Post> findByMember(Member member);
 }
