+++ src/main/java/com/example/miniproject/service/MemberService.java
@@ -2,6 +2,9 @@ package com.example.miniproject.service;
 
 import com.example.miniproject.domain.Member;
 import com.example.miniproject.dto.*;
+import com.example.miniproject.dto.request.LoginRequestDto;
+import com.example.miniproject.dto.request.MemberRequestDto;
+import com.example.miniproject.dto.response.MemberResponseDto;
 import com.example.miniproject.jwt.TokenProvider;
 import com.example.miniproject.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
